refactor(hooks): move custom hooks to src/hooks for clarity\n\n- Move usePdfTextExtractor and useQuizGenerator to src/hooks\n- Update imports in main page\n- Improve project structure and maintainability
